+++ __tests__/browser/global-setup.js
@@ -1,0 +1,6 @@
+import 'babel-polyfill';
+import { setup as setupPuppeteer } from 'jest-environment-puppeteer';
+
+module.exports = async function globalSetup(globalConfig) {
+  await setupPuppeteer(globalConfig);
+};