+++ index.js
@@ -6,9 +6,10 @@ const unified = require('unified');
 
 /* Unified Plugins
  */
-const sanitize = require('hast-util-sanitize/lib/github.json');
+const sanitize = require('./sanitize.schema');
 
 const generateTOC = require('mdast-util-toc');
+
 const mapNodes = require('unist-util-map');
 const { selectAll } = require('unist-util-select');
 
@@ -19,6 +20,7 @@ const remarkParse = require('remark-parse');
 const remarkStringify = require('remark-stringify');
 const remarkBreaks = require('remark-breaks');
 const remarkSlug = require('remark-slug');
+const remarkFrontmatter = require('remark-frontmatter');
 
 // rehype plugins
 const rehypeSanitize = require('rehype-sanitize');
@@ -60,43 +62,6 @@ const toPlainText = require('./processor/plugin/plain-text');
 
 // Processor Option Defaults
 const options = require('./options.json');
-
-// Sanitization Schema Defaults
-sanitize.clobberPrefix = '';
-
-sanitize.tagNames.push('span', 'style');
-sanitize.attributes['*'].push('class', 'className', 'align', 'style');
-
-/**
- * @todo don't manually whitelist custom component attributes
- *       within the engine!
- * @todo change `link` to `href`
- */
-sanitize.attributes['tutorial-tile'] = ['backgroundColor', 'emoji', 'link'];
-
-sanitize.tagNames.push('rdme-pin');
-
-sanitize.tagNames.push('rdme-embed');
-sanitize.attributes['rdme-embed'] = [
-  'url',
-  'provider',
-  'html',
-  'title',
-  'href',
-  'iframe',
-  'width',
-  'height',
-  'image',
-  'favicon',
-];
-
-sanitize.attributes.a = ['href', 'title', 'class', 'className', 'download'];
-
-sanitize.tagNames.push('figure');
-sanitize.tagNames.push('figcaption');
-
-sanitize.tagNames.push('input'); // allow GitHub-style todo lists
-sanitize.ancestors.input = ['li'];
 
 /**
  * Normalize Magic Block Raw Text
@@ -147,6 +112,7 @@ export function processor(opts = {}) {
    */
   return unified()
     .use(remarkParse, opts.markdownOptions)
+    .use(remarkFrontmatter, ['yaml', 'toml'])
     .data('settings', opts.settings)
     .data('compatibilityMode', opts.compatibilityMode)
     .use(!opts.correctnewlines ? remarkBreaks : () => {})
@@ -227,6 +193,8 @@ export function reactTOC(tree, opts = {}) {
     createElement: React.createElement,
     components: {
       p: React.Fragment,
+      'readme-variable': Variable,
+      'readme-glossary-item': GlossaryItem,
     },
   });