+++ webpack.config.js
@@ -1,16 +1,11 @@
 const path = require('path');
-const merge = require('webpack-merge');
-const ExtractCSS = require('mini-css-extract-plugin');
+const { merge } = require('webpack-merge');
 const TerserPlugin = require('terser-webpack-plugin');
 
-const output = {
-  path: path.resolve(__dirname, 'dist'),
-  filename: '[name].js',
-  libraryTarget: 'commonjs2',
-};
+const common = require('./webpack.common');
 
-const browserConfig = {
-  entry: ['./index.js'],
+const browserConfig = merge(common, {
+  entry: './index.js',
   externals: {
     '@readme/variable': '@readme/variable',
     react: {
@@ -28,70 +23,16 @@ const browserConfig = {
       umd: 'react-dom',
     },
   },
-  output,
-  plugins: [
-    new ExtractCSS({
-      filename: '[name].css',
-    }),
-  ],
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: '[name].js',
+    libraryTarget: 'commonjs2',
+  },
   optimization: {
     minimize: false,
     minimizer: [new TerserPlugin()],
-    // concatenateModules: false,
-    // namedModules: true,
-    // namedChunks: true,
-    // removeAvailableModules: false,
-    // flagIncludedChunks: false,
-    // occurrenceOrder: false,
   },
-  module: {
-    rules: [
-      {
-        test: /node_modules\/.*(is-plain-obj|parse5)\/.*.js$/,
-        use: {
-          loader: 'babel-loader',
-          options: { extends: './.babelrc' },
-        },
-      },
-      {
-        test: /\.jsx?$/,
-        exclude: /node_modules\/(?!@readme\/[\w-]+\/)/,
-        use: {
-          loader: 'babel-loader',
-          options: { extends: './.babelrc' },
-        },
-      },
-      {
-        test: /\.css$/,
-        loaders: [ExtractCSS.loader, 'css-loader'],
-      },
-      {
-        test: /\.scss$/,
-        loaders: [ExtractCSS.loader, 'css-loader', 'sass-loader'],
-      },
-      {
-        // eslint-disable-next-line unicorn/no-unsafe-regex
-        test: /\.(ttf|eot|svg|woff(2)?)(\?[a-z0-9=&.]+)?$/,
-        loader: 'file-loader?name=dist/fonts/[hash].[ext]',
-        exclude: /(node_modules)/,
-      },
-      {
-        test: /\.(txt|md)$/i,
-        use: 'raw-loader',
-      },
-    ],
-  },
-  resolve: {
-    extensions: ['.js', '.json', '.jsx'],
-  },
-  devServer: {
-    port: 6699,
-    contentBase: './dist',
-    watchContentBase: true,
-    compress: true,
-    // hot: true,
-  },
-};
+});
 
 const serverConfig = merge(browserConfig, {
   target: 'node',